Added 'barenco' mode to mcx
Added a new mode for the mcx method of the ExtendedQuantumCircuit that uses the Barenco MCMT implementation.

In the already implemented methods the target Qubit is always the most significant one, contrary to the implementation I make where the least significant Qubit is the target.
For now I will make a distinction between old methods and new methods when appending them to circumvent this issue.